Simplified cache with only complete postings.
